Add configuration system and enhanced controller functionality
- Add ConfigManager for persistent settings
- Redesign button layout with comprehensive controls:
  * Left shoulder: Win+Tab (Task View)
  * Right shoulder: Enter key
  * Back: Escape, Guide: Windows key
  * Stick clicks: Middle mouse and Screenshot
  * Triggers: Media previous/next track
- Add mouse and scroll sensitivity adjustment via D-pad
- Add right stick Y-axis inversion support
- Improve button state tracking with proper press/release detection
- Update Windows package with latest executable and documentation

+++ include/media_controller.h
@@ -27,7 +27,7 @@ private:
     bool is_initialized_;
     
 #ifdef _WIN32
-    void sendMediaKey(DWORD key);
+    void sendMediaKey(unsigned long key);
 #elif __linux__
     void sendMediaCommand(const char* command);
 #elif __APPLE__
